:recycle: refactor(BTS-300): 중복 제거 및 레이아웃, 스토어 레이어 변경
