+++ src/app/(home)/page.tsx
@@ -1,5 +1,5 @@
 import { Session1, Session2, Session3, Session4 } from '@/features/home';
-import { Footer } from '@/shared/components/layout/footer';
+import { Footer } from '@/layout/footer';
 
 export default function home() {
   return (
