+++ src/app/(private)/dashboard/study-note/[id]/page.tsx
@@ -1,6 +1,6 @@
 import StudyNoteDetailContentsSection from '@/features/dashboard/studynote/detail/components/contents-section';
 import StudyNoteDetailMetaSection from '@/features/dashboard/studynote/detail/components/meta-section';
-import { ColumnLayout } from '@/shared/components/layout/column-layout';
+import { ColumnLayout } from '@/layout/column-layout';
 
 export default async function StudyNotePage({
   params,
