+++ src/app/(private)/study-rooms/[id]/(study-note)/qna/[contextId]/page.tsx
@@ -1,5 +1,5 @@
 import { QuestionDetail } from '@/features/qna/components/detail/qna-detail';
-import { ColumnLayout } from '@/shared/components/layout/column-layout';
+import { ColumnLayout } from '@/layout/column-layout';
 
 type Props = {
   params: Promise<{ id: string; contextId: string }>;
