+++ src/app/(private)/study-rooms/[id]/(study-note)/qna/new/page.tsx
@@ -1,5 +1,5 @@
 import QnAWriteDetail from '@/features/qna/components/write/qna-write-detail';
-import { ColumnLayout } from '@/shared/components/layout/column-layout';
+import { ColumnLayout } from '@/layout/column-layout';
 
 type Props = {
   params: Promise<{ id: string }>;
