+++ src/features/dashboard/studynote/detail/components/contents-section.tsx
@@ -6,8 +6,8 @@ import Image from 'next/image';
 
 import EllipsisIcon from '@/assets/icons/ellipsis-vertical.svg';
 import { useStudyNoteDetailQuery } from '@/features/dashboard/studynote/detail/service/query';
+import { ColumnLayout } from '@/layout/column-layout';
 import { TextViewer } from '@/shared/components/editor';
-import { ColumnLayout } from '@/shared/components/layout/column-layout';
 import { format } from 'date-fns';
 import { ko } from 'date-fns/locale';
