+++ src/features/qna/components/detail/qna-detail.tsx
@@ -4,7 +4,7 @@ import { Key } from 'react';
 
 import Image from 'next/image';
 
-import { ColumnLayout } from '@/shared/components/layout/column-layout';
+import { ColumnLayout } from '@/layout/column-layout';
 import { useRole } from '@/shared/hooks/use-role';
 import { cn } from '@/shared/lib/utils';
