+++ src/features/qna/components/write/qna-message-write-area.tsx
@@ -5,8 +5,8 @@ import { Controller, useFormContext } from 'react-hook-form';
 
 import { useRouter } from 'next/navigation';
 
+import { ColumnLayout } from '@/layout/column-layout';
 import { TextEditor } from '@/shared/components/editor';
-import { ColumnLayout } from '@/shared/components/layout/column-layout';
 import { Button } from '@/shared/components/ui/button';
 import { Form } from '@/shared/components/ui/form';
 import { ROUTE } from '@/shared/constants/route';
