+++ src/layout/index.ts
@@ -1,0 +1,4 @@
+export { Header } from './header';
+export { Footer } from './footer';
+export { Sidebar } from './sidebar';
+export { ColumnLayout } from './column-layout';
