--- src/shared/common/index.ts
@@ -1,1 +1,0 @@
-export * from './query.make-search-option';
