+++ src/shared/constants/route.ts
@@ -1,4 +1,6 @@
-// 프론트엔드 내비게이션 경로 관리
+/* ─────────────────────────────────────────────────────
+ * PUBLIC
+ * ────────────────────────────────────────────────────*/
 export const ROUTE = {
   HOME: '/',
   LOGIN: '/login',
